review of Simon
